MOO-487 Fix section double loading
This commit moves the logic to generate the form links in LeftMenuFragmentController: for each form, the controller checks if there is a previous existing encounter and generate the edit URL (or the 'enter new' URL if there is no previous encounter). This resolves the double-load issue and simplifies the xml forms.